requirements.txt
altair==5.3.0
annotated-types==0.6.0
blinker==1.7.0
cachetools==5.3.3
click==8.1.7
distro==1.9.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.17.0
pandas==2.2.1
pillow==10.3.0
pinecone-client==3.2.2
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
regex==2023.12.25
rich==13.7.1
safetensors==0.4.2
smmap==5.0.1
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
tzdata==2024.1
webencodings==0.5.1
